Split up test files into categories
